add problem desc functions

+++ include/miopen/miopen.h
@@ -6613,7 +6613,7 @@ MIOPEN_EXPORT miopenStatus_t miopenNLLLossForward(miopenHandle_t handle,
                                                   const void* weight,
                                                   const miopenTensorDescriptor_t outputDesc,
                                                   void* output,
-                                                  long ignore_index);
+                                                  int ignore_index);
 /** @} */
 // CLOSEOUT nllloss DOXYGEN GROUP
 #endif // MIOPEN_BETA_API
